feat(trainingAndQualsHowToTab): Move helper functions into excelUtils file
